+++ 1.21.11/Common/src/main/java/fuzs/resourcepackoverrides/server/packs/PackSelectionOverride.java
@@ -9,6 +9,8 @@ import net.minecraft.server.packs.repository.Pack;
 import net.minecraft.server.packs.repository.PackCompatibility;
 import org.jspecify.annotations.Nullable;
 
+import java.util.Objects;
+
 public record PackSelectionOverride(@Nullable Component title,
                                     @Nullable Component description,
                                     Pack.@Nullable Position defaultPosition,
@@ -16,7 +18,6 @@ public record PackSelectionOverride(@Nullable Component title,
                                     @Nullable Boolean fixedPosition,
                                     @Nullable Boolean required,
                                     @Nullable Boolean hidden) {
-
     public static final PackSelectionOverride EMPTY = new PackSelectionOverride(null,
             null,
             null,
@@ -30,23 +31,27 @@ public record PackSelectionOverride(@Nullable Component title,
     }
 
     public void apply(Pack pack) {
-        if (this.title != null) {
+        if (this.title != null && !Objects.equals(this.title, pack.getTitle())) {
             pack.location = new PackLocationInfo(pack.location.id(),
                     this.title,
                     pack.location.source(),
                     pack.location.knownPackInfo());
         }
 
-        if (this.required() != null || this.fixedPosition() != null || this.defaultPosition() != null) {
-            boolean required = this.required != null ? this.required : pack.selectionConfig.required();
+        if (this.required != null && !Objects.equals(this.required, pack.isRequired())
+                || this.defaultPosition != null && !Objects.equals(this.defaultPosition, pack.getDefaultPosition())
+                || this.fixedPosition != null && !Objects.equals(this.fixedPosition, pack.isFixedPosition())) {
+            boolean required = this.required != null ? this.required : pack.isRequired();
             Pack.Position defaultPosition =
-                    this.defaultPosition != null ? this.defaultPosition : pack.selectionConfig.defaultPosition();
-            boolean fixedPosition =
-                    this.fixedPosition != null ? this.fixedPosition : pack.selectionConfig.fixedPosition();
+                    this.defaultPosition != null ? this.defaultPosition : pack.getDefaultPosition();
+            boolean fixedPosition = this.fixedPosition != null ? this.fixedPosition : pack.isFixedPosition();
             pack.selectionConfig = new PackSelectionConfig(required, defaultPosition, fixedPosition);
         }
 
-        if (this.description() != null || this.compatibility() != null || this.hidden() != null) {
+        if (this.description != null && !Objects.equals(this.description, pack.getDescription())
+                || this.compatibility != null && !Objects.equals(this.compatibility, pack.getCompatibility())
+                || this.hidden != null && !Objects.equals(this.hidden,
+                ClientAbstractions.INSTANCE.isPackHidden(pack))) {
             Component description = this.description != null ? this.description : pack.getDescription();
             PackCompatibility compatibility =
                     this.compatibility != null ? this.compatibility() : pack.getCompatibility();