+++ 1.21.11/Common/src/main/java/fuzs/resourcepackoverrides/services/ClientAbstractions.java
@@ -17,9 +17,9 @@ public interface ClientAbstractions {
 
     boolean isPackHidden(Pack pack);
 
-    void setPackHidden(Pack pack, boolean hidden);
+    void setPackHidden(Pack pack, boolean isHidden);
 
-    Pack.Metadata createPackInfo(Component description, PackCompatibility compatibility, FeatureFlagSet features, List<String> overlays, boolean isHidden);
+    Pack.Metadata createPackInfo(Component descriptionComponent, PackCompatibility packCompatibility, FeatureFlagSet featureFlagSet, List<String> overlays, boolean isHidden);
 
     enum ModLoader {
         FABRIC, NEOFORGE