+++ 1.21.10/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.10.1-1.21.10] - 2025-12-26
+
+### Changed
+
+- When generating a fresh new config for the mod, all resource packs are made to show up as compatible by default
+
+### Fixed
+
+- Hidden packs are no longer always enabled automatically on Fabric
+
 ## [v21.10.0-1.21.10] - 2025-12-16
 
 ### Changed